server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
SQLAlchemy==1.4.47
pydantic==1.10.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==0.20.0
alembic==1.10.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-jose==3.3.0
passlib==1.7.4
jose==1.0.0
